style fixes
instead of making that function a one-liner, we put it back to
multiple lines and instead wrap it in a echo call to work around the
race condition from the outside.

i haven't found such race conditions so far, so i am not clear on how
to reproduce this, but it should still work while making the code
cleaner.

At the very least, this doesn't seem to have problems:

    ./battery-status-collect & ./battery-status-collect